requirements.txt
fastapi==0.115.0
uvicorn==0.30.1

SQLAlchemy==2.0.30
alembic==1.13.1

openai==1.45.0
chromadb==0.5.3

apscheduler==3.10.4

jinja2==3.1.4
python-multipart==0.0.9

python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.0

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

pytest==8.2.2
httpx==0.27.0
alembic>=1.13.0

